Serving an attack classifier with django/requirements.txt
Django==3.1
djangorestframework==3.11.1
numpy==1.16.4
tensorflow==1.14.0
Unidecode==1.1.1


